fix(watch): open dictation directly from the Chat microphone button
The round voice button in Apple Watch Chat looked like "talk" but opened the
same keyboard/Scribble input picker as the Message OpenClaw pill; only the
spoken reply differed. Apple ships no Speech framework on watchOS, so the
system dictation screen is the on-watch transcript. WatchKit opens it directly
only for a nil suggestion list in plain mode, which the composer now models as
an explicit input kind: the microphone (mic.fill) starts dictation and requests
the spoken reply, the pill (keyboard) keeps typed input silent. The submitted
text now shows in Chat as a pending bubble until the iPhone snapshot echoes it,
and the dictation path plays the HIG start/success/failure haptics.

The chat timeline and composer move to WatchChatTimelineView.swift because
WatchInboxView.swift crossed the SwiftLint file_length cap; the move is
otherwise verbatim.

Related: #100224, #135765
