+++ apps/ios/WatchApp/Sources/WatchInboxView.swift
@@ -679,7 +679,7 @@ private enum WatchAvatarSource {
     }
 }
 
-private struct WatchClawAvatar: View {
+struct WatchClawAvatar: View {
     var size: CGFloat
     var imageSource: String?
     var text: String?
@@ -945,7 +945,7 @@ private struct WatchSecondaryLabel: View {
     }
 }
 
-private struct WatchSecondaryButton: View {
+struct WatchSecondaryButton: View {
     let title: LocalizedStringKey
     let action: () -> Void
 
@@ -1050,7 +1050,7 @@ private struct WatchDecisionButton: View {
     }
 }
 
-private struct WatchTinyStatus: View {
+struct WatchTinyStatus: View {
     let text: String
 
     var body: some View {
@@ -1115,344 +1115,6 @@ private enum WatchExecApprovalDisplay {
             return statusText
         }
         return record.isResolving ? String(localized: "Sending decision...") : nil
-    }
-}
-
-private struct WatchChatBubble: View {
-    let item: WatchChatItem
-    var avatarImageSource: String?
-    var avatarText: String?
-
-    var body: some View {
-        HStack(alignment: .bottom, spacing: 6) {
-            if !self.isUser {
-                WatchClawAvatar(
-                    size: 18,
-                    imageSource: self.avatarImageSource,
-                    text: self.avatarText)
-            } else {
-                Spacer(minLength: 20)
-            }
-
-            VStack(alignment: self.isUser ? .trailing : .leading, spacing: 3) {
-                Text(self.roleTitle)
-                    .font(WatchClawType.label(size: 9, weight: .bold))
-                    .foregroundStyle(self.isUser ? .secondary : WatchClawStyle.accent)
-                Text(self.item.text)
-                    .font(WatchClawType.body(size: 13))
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            .padding(.horizontal, 9)
-            .padding(.vertical, 7)
-            .frame(maxWidth: 132, alignment: self.isUser ? .trailing : .leading)
-            .background {
-                RoundedRectangle(cornerRadius: 15, style: .continuous)
-                    .fill(self.isUser ? WatchClawStyle.accent.opacity(0.88) : WatchClawStyle.surface)
-            }
-
-            if self.isUser {
-                WatchMiniUserDot()
-            } else {
-                Spacer(minLength: 20)
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-
-    private var isUser: Bool {
-        self.item.role.lowercased() == "user"
-    }
-
-    private var roleTitle: String {
-        switch self.item.role.lowercased() {
-        case "user":
-            String(localized: "You")
-        case "system":
-            String(localized: "System")
-        default:
-            "OpenClaw"
-        }
-    }
-}
-
-private struct WatchChatTimelineView: View {
-    @Environment(\.scenePhase) private var scenePhase
-    var store: WatchInboxStore
-    let items: [WatchChatItem]
-    let statusText: String
-    let sendStatusText: String?
-    var avatarImageSource: String?
-    var avatarText: String?
-    var onRefresh: (() -> Void)?
-    var onSendMessage: ((String) -> String?)?
-    @State private var speechPlayback = WatchSpeechPlayback()
-    @State private var voiceReplyTimeout: Task<Void, Never>?
-    @State private var isVisible = false
-
-    var body: some View {
-        VStack(spacing: 7) {
-            ScrollView {
-                VStack(alignment: .leading, spacing: 8) {
-                    if self.items.isEmpty {
-                        WatchChatEmptyState(statusText: self.statusText)
-                    } else {
-                        ForEach(self.items) { item in
-                            WatchChatBubble(
-                                item: item,
-                                avatarImageSource: self.avatarImageSource,
-                                avatarText: self.avatarText)
-                        }
-                    }
-
-                    if let sendStatusText, !sendStatusText.isEmpty {
-                        WatchTinyStatus(text: sendStatusText)
-                    }
-
-                    if let voiceStatusText = self.voiceStatusText {
-                        VStack(alignment: .leading, spacing: 3) {
-                            // Watch TTS runs through AVSpeechSynthesizer, which has no
-                            // metering API, so speaking uses the wave's synthetic pulse.
-                            TalkWaveformView(
-                                phase: self.speechPlayback.isSpeaking ? .speaking(level: nil) : .thinking)
-                                .frame(height: 24)
-                                .accessibilityHidden(true)
-                            WatchTinyStatus(text: voiceStatusText)
-                        }
-                    }
-
-                    if let errorText = self.speechPlayback.errorText {
-                        WatchTinyStatus(text: errorText)
-                    }
-
-                    WatchSecondaryButton(title: "Refresh") {
-                        self.onRefresh?()
-                    }
-                }
-                .padding(.horizontal, 8)
-                .padding(.top, 8)
-                .padding(.bottom, 4)
-                .frame(maxWidth: .infinity, alignment: .topLeading)
-            }
-            .scrollIndicators(.hidden)
-
-            WatchChatComposer(
-                onComposeMessage: {
-                    self.presentMessageInput(spokenReply: false)
-                },
-                onStartVoiceTurn: {
-                    self.presentMessageInput(spokenReply: true)
-                },
-                isAwaitingVoiceReply: self.store.isAwaitingVoiceReply,
-                onCancelVoiceTurn: {
-                    self.cancelVoiceTurn()
-                },
-                isSpeaking: self.speechPlayback.isSpeaking,
-                onStopSpeaking: {
-                    self.speechPlayback.stop()
-                })
-                .padding(.horizontal, 7)
-                .padding(.bottom, 5)
-        }
-        .background(WatchClawStyle.background.ignoresSafeArea())
-        .navigationTitle("Chat")
-        .onChange(of: self.store.chatCompletion) { _, _ in
-            self.handleCompletedVoiceTurn()
-        }
-        .onChange(of: self.scenePhase) { _, phase in
-            if phase == .active {
-                self.handleCompletedVoiceTurn()
-                self.scheduleVoiceReplyTimeout()
-            } else if phase == .background {
-                self.voiceReplyTimeout?.cancel()
-                self.speechPlayback.stop()
-            }
-        }
-        .onAppear {
-            self.isVisible = true
-            self.handleCompletedVoiceTurn()
-            self.scheduleVoiceReplyTimeout()
-        }
-        .onDisappear {
-            self.isVisible = false
-            self.voiceReplyTimeout?.cancel()
-            self.speechPlayback.stop()
-        }
-    }
-
-    private var voiceStatusText: String? {
-        if self.speechPlayback.isSpeaking {
-            return String(localized: "Speaking reply…")
-        }
-        if self.store.isAwaitingVoiceReply {
-            return String(localized: "Waiting for spoken reply…")
-        }
-        return nil
-    }
-
-    private func presentMessageInput(spokenReply: Bool) {
-        let chatSession = self.store.appSnapshot?.chatSessionIdentity
-        self.speechPlayback.stop()
-        WatchNativeTextInput.present(suggestions: []) { text in
-            guard self.store.appSnapshot?.chatSessionIdentity == chatSession else {
-                self.store.markAppCommandBlocked(
-                    .sendChat,
-                    reason: String(localized: "Chat changed on iPhone. Your message was not sent."))
-                return
-            }
-            guard let commandId = self.onSendMessage?(text) else { return }
-            if spokenReply {
-                self.store.beginVoiceTurn(commandId: commandId)
-                self.scheduleVoiceReplyTimeout()
-            }
-        }
-    }
-
-    private func handleCompletedVoiceTurn() {
-        guard self.isVisible, self.scenePhase == .active,
-              let reply = self.store.takeVoiceReply()
-        else { return }
-        self.voiceReplyTimeout?.cancel()
-        self.speechPlayback.speak(reply)
-    }
-
-    private func cancelVoiceTurn() {
-        self.voiceReplyTimeout?.cancel()
-        self.store.cancelVoiceTurn()
-    }
-
-    private func scheduleVoiceReplyTimeout() {
-        self.voiceReplyTimeout?.cancel()
-        guard let delayNanoseconds = self.store.voiceReplyTimeoutNanoseconds() else { return }
-        self.voiceReplyTimeout = Task { @MainActor in
-            try? await Task.sleep(nanoseconds: delayNanoseconds)
-            guard !Task.isCancelled else { return }
-            self.store.cancelVoiceTurn()
-        }
-    }
-}
-
-private struct WatchChatEmptyState: View {
-    let statusText: String
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 4) {
-            Text("No chat synced")
-                .font(WatchClawType.title(size: 16))
-                .lineLimit(2)
-            Text(self.statusText)
-                .font(WatchClawType.body(size: 12))
-                .foregroundStyle(.secondary)
-                .lineLimit(3)
-            Text("Tap the message pill below to start from your watch.")
-                .font(WatchClawType.body(size: 11, weight: .medium))
-                .foregroundStyle(WatchClawStyle.accent)
-                .lineLimit(2)
-        }
-        .padding(.horizontal, 10)
-        .padding(.vertical, 9)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background {
-            RoundedRectangle(cornerRadius: 17, style: .continuous)
-                .fill(WatchClawStyle.surface)
-                .overlay {
-                    RoundedRectangle(cornerRadius: 17, style: .continuous)
-                        .strokeBorder(WatchClawStyle.border, lineWidth: 1)
-                }
-        }
-    }
-}
-
-private struct WatchMiniUserDot: View {
-    var body: some View {
-        Text("You")
-            .font(WatchClawType.label(size: 8, weight: .bold))
-            .foregroundStyle(.white.opacity(0.86))
-            .frame(width: 22, height: 18)
-            .background {
-                Capsule(style: .continuous)
-                    .fill(Color.white.opacity(0.10))
-            }
-    }
-}
-
-private struct WatchChatComposer: View {
-    let onComposeMessage: () -> Void
-    let onStartVoiceTurn: () -> Void
-    let isAwaitingVoiceReply: Bool
-    let onCancelVoiceTurn: () -> Void
-    let isSpeaking: Bool
-    let onStopSpeaking: () -> Void
-
-    var body: some View {
-        HStack(spacing: 6) {
-            Button {
-                self.onComposeMessage()
-            } label: {
-                HStack(spacing: 5) {
-                    Text("Message OpenClaw")
-                        .font(WatchClawType.body(size: 12, weight: .semibold))
-                        .foregroundStyle(.secondary)
-                        .lineLimit(1)
-                    Spacer(minLength: 0)
-                    Image(systemName: "text.bubble")
-                        .font(WatchClawType.symbol(size: 12, weight: .semibold))
-                        .foregroundStyle(.secondary)
-                }
-                .padding(.horizontal, 11)
-                .padding(.vertical, 10)
-                .background {
-                    Capsule(style: .continuous)
-                        .fill(Color.white.opacity(0.09))
-                        .overlay {
-                            Capsule(style: .continuous)
-                                .strokeBorder(Color.white.opacity(0.16), lineWidth: 1)
-                        }
-                }
-            }
-            .buttonStyle(.plain)
-            .disabled(self.isAwaitingVoiceReply)
-
-            Button {
-                if self.isSpeaking {
-                    self.onStopSpeaking()
-                } else if self.isAwaitingVoiceReply {
-                    self.onCancelVoiceTurn()
-                } else {
-                    self.onStartVoiceTurn()
-                }
-            } label: {
-                Group {
-                    if self.isSpeaking {
-                        Image(systemName: "speaker.slash.fill")
-                            .font(WatchClawType.symbol(size: 13, weight: .bold))
-                    } else if self.isAwaitingVoiceReply {
-                        Image(systemName: "xmark")
-                            .font(WatchClawType.symbol(size: 13, weight: .bold))
-                    } else {
-                        WatchVoiceGlyph()
-                    }
-                }
-                .foregroundStyle(.white)
-                .frame(width: 20, height: 20)
-                .padding(8)
-                .background {
-                    Circle()
-                        .fill(WatchClawStyle.hotGradient)
-                }
-            }
-            .buttonStyle(.plain)
-            .accessibilityLabel(self.voiceButtonAccessibilityLabel)
-        }
-    }
-
-    private var voiceButtonAccessibilityLabel: String {
-        if self.isSpeaking {
-            return String(localized: "Stop speaking")
-        }
-        if self.isAwaitingVoiceReply {
-            return String(localized: "Cancel spoken reply")
-        }
-        return String(localized: "Start voice turn")
     }
 }
