+++ apps/ios/WatchApp/Sources/WatchNativeTextInput.swift
@@ -2,14 +2,14 @@ import Foundation
 import WatchKit
 
 enum WatchNativeTextInput {
+    /// WatchKit opens the system dictation screen directly only for a nil
+    /// suggestion list in `.plain` mode; any list or emoji mode shows the
+    /// keyboard/Scribble/dictation picker first.
     @MainActor
-    static func present(
-        suggestions: [String],
-        onSubmit: @escaping (String) -> Void)
-    {
+    static func present(_ kind: WatchComposerInputKind, onSubmit: @escaping (String) -> Void) {
         WKApplication.shared().visibleInterfaceController?.presentTextInputController(
-            withSuggestions: suggestions,
-            allowedInputMode: .allowEmoji)
+            withSuggestions: kind.suggestions,
+            allowedInputMode: kind.allowsEmoji ? .allowEmoji : .plain)
         { results in
             guard let text = results?.compactMap(stringValue).first?
                 .trimmingCharacters(in: .whitespacesAndNewlines),
